refactor front-end refresh token logic to simplify
