+++ application/react/App/App.tsx
@@ -24,7 +24,6 @@ import { ProtectedRoute } from '../_auth/ProtectedRoute/ProtectedRoute';
 import { Login } from '../_auth/Login/Login';
 import { USER, ADMIN } from '../../api/enums/userTypesEnum'
 import { Home } from '../Home/Home';
-import { PersistLogin } from '../_auth/PersistLogin/PersistLogin'
 
 const queryClient = new QueryClient();
 
@@ -33,6 +32,7 @@ export function App() {
     <QueryClientProvider client={queryClient}>
       <Routes >
         <Route path='react-ui'>
+
           {/* PUBLIC Routes without nav */}
           <Route path='login' element={<Login />}></Route>
 
@@ -41,33 +41,31 @@ export function App() {
             <Route path='*' element={<PageNotFound />} /> {/* TODO @Storm: Build this page */}
             
             {/* PROTECTED ROUTES: USER / ADMIN */}
-            <Route element={<PersistLogin />}>
-              <Route element={<ProtectedRoute allowedRoles={[USER, ADMIN]}/>}>
-                
-                <Route path='' element={<Home />}></Route>
-                <Route path='inventory' element={<Inventory />}></Route>
+            <Route element={<ProtectedRoute allowedRoles={[USER, ADMIN]}/>}>
+              
+              <Route path='' element={<Home />}></Route>
+              <Route path='inventory' element={<Inventory />}></Route>
 
-                <Route path='forms'>
-                  <Route path='material-length-adjustment' element={<MaterialLengthAdjustmentForm />} />
-                  <Route path='delivery-method/:mongooseId?' element={<DeliveryMethodForm />} />
-                  <Route path='credit-term/:mongooseId?' element={<CreditTermForm />} />
-                  <Route path='quote' element={<QuoteForm />} />
-                  <Route path='customer/:mongooseId?' element={<CustomerForm />} />
-                  <Route path="liner-type/:mongooseId?" element={<LinerTypeForm />} /> {/* TODO (6-5-2024): Enforce admin routes only render for admins */}
-                  <Route path='material/:mongooseId?' element={<MaterialForm />} />
-                  <Route path='adhesive-category/:mongooseId?' element={<AdhesiveCategoryForm />} />
-                  <Route path='material-order/:mongooseId?' element={<MaterialOrderForm />} />
-                </Route>
+              <Route path='forms'>
+                <Route path='material-length-adjustment' element={<MaterialLengthAdjustmentForm />} />
+                <Route path='delivery-method/:mongooseId?' element={<DeliveryMethodForm />} />
+                <Route path='credit-term/:mongooseId?' element={<CreditTermForm />} />
+                <Route path='quote' element={<QuoteForm />} />
+                <Route path='customer/:mongooseId?' element={<CustomerForm />} />
+                <Route path="liner-type/:mongooseId?" element={<LinerTypeForm />} /> {/* TODO (6-5-2024): Enforce admin routes only render for admins */}
+                <Route path='material/:mongooseId?' element={<MaterialForm />} />
+                <Route path='adhesive-category/:mongooseId?' element={<AdhesiveCategoryForm />} />
+                <Route path='material-order/:mongooseId?' element={<MaterialOrderForm />} />
+              </Route>
 
-                <Route path='tables'>
-                  <Route path='credit-term' element={<CreditTermTable />} />
-                  <Route path='delivery-method' element={<DeliveryMethodTable />} />
-                  <Route path='liner-type' element={<LinerTypeTable />} />
-                  <Route path='material' element={<MaterialTable />} />
-                  <Route path='adhesive-category' element={<AdhesiveCategoryTable />} />
-                  <Route path='customer' element={<CustomerTable />} />
-                  <Route path='material-order' element={<MaterialOrderTable />} />
-                </Route>
+              <Route path='tables'>
+                <Route path='credit-term' element={<CreditTermTable />} />
+                <Route path='delivery-method' element={<DeliveryMethodTable />} />
+                <Route path='liner-type' element={<LinerTypeTable />} />
+                <Route path='material' element={<MaterialTable />} />
+                <Route path='adhesive-category' element={<AdhesiveCategoryTable />} />
+                <Route path='customer' element={<CustomerTable />} />
+                <Route path='material-order' element={<MaterialOrderTable />} />
               </Route>
             </Route>
           </Route>
