+++ application/react/Navbars/HorizontalNavbar/HorizontalNavbar.tsx
@@ -14,7 +14,7 @@ export const HorizontalNavbar = () => {
               <li className="navbar-links nav-dropdown-trigger" id="insights-dropdown-trigger">
                   <div className="insights flex-center-center-row">
                       <i className="fa-regular fa-chart-network"></i>Insights
-                      <svg xmlns="http://www.w3.org/2000/svg" width="22px" height="22px" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" className="feather feather-chevron-down"><polyline points="6 9 12 15 18 9"></polyline></svg>
+                      <svg xmlns="http://www.w3.org/2000/svg" width="22px" height="22px" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-chevron-down"><polyline points="6 9 12 15 18 9"></polyline></svg>
                   </div>
                   <ul className="dropdown-menu insights-dropdown">
                       <li><a href="/recipes/create"><i className="fa-solid fa-list-timeline"></i>Master Log</a></li>
